patch: move pebble observer to common
